+++ src/gr-pulsed_power/lib/integration_impl.h
@@ -23,16 +23,16 @@ private:
     float d_last_value;
     time_point<system_clock> d_last_save;
     time_point<system_clock> d_last_reset;
-    float d_sum;
+    double d_sum;
     int d_duration;
     const std::string d_filename;
 
     int get_values_from_file(time_point<system_clock>& last_reset,
                              time_point<system_clock>& last_save,
-                             float& sum);
+                             double& sum);
     int write_to_file(time_point<system_clock> last_reset,
                       time_point<system_clock> last_save,
-                      float sum);
+                      double sum);
 
 public:
     integration_impl(int decimation,
@@ -43,7 +43,7 @@ public:
 
     void add_new_steps(float* out, const float* sample, int noutput_items) override;
 
-    void integrate(float& out,
+    void integrate(double& out,
                    const float* sample,
                    int n_samples,
                    bool calculate_with_last_value) override;
